requirements.txt
APScheduler==3.6.3
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
h11==0.13.0
idna==3.3
outcome==1.1.0
pycodestyle==2.8.0
pycparser==2.21
pyOpenSSL==22.0.0
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
toml==0.10.2
tornado==6.1
trio==0.19.0
trio-websocket==0.9.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
webdriver-manager==3.5.3
wsproto==1.0.0
